Reimplement loading jnlp file with separate window so client is being initiated once. Hopefully it will helt the GrotArmy with issue he is having with running client under linux system.
